[Core] Remove the ArrayInput processor
For an URL this format is extreme verbose and JSON pretty much provides the exact same
functionality (but in a much nicer syntax).
