+++ lib/Core/Tests/Loader/ConditionExporterLoaderTest.php
@@ -47,7 +47,6 @@ final class ConditionExporterLoaderTest extends TestCase
     {
         return [
             ['json', Exporter\JsonExporter::class],
-            ['array', Exporter\ArrayExporter::class],
             ['string_query', Exporter\StringQueryExporter::class],
             ['norm_string_query', Exporter\NormStringQueryExporter::class],
         ];
